Refactor Private Marketplace access validation into a provider strategy
Extract all eligibility/access-validation logic out of
WorkbenchExtensionGalleryManifestService into a dedicated, provider-agnostic
ExtensionGalleryAccessValidator, and split the GitHub-vs-Microsoft branching
into IExtensionGalleryAccessProvider strategy classes. This debloats the host
service (it now only builds a status sink and delegates) and isolates each
identity system's account resolution + eligibility check.

Replace the hand-rolled monotonic validationEpoch TOCTOU counter with a
CancellationTokenSource held in a MutableDisposable: assigning a new source
cancels/disposes the prior one, and each validation re-checks
token.isCancellationRequested immediately before mutating status/cache/manifest,
so a superseded in-flight validation cannot commit a stale verdict for an
account that is no longer current. Addresses reviewer feedback that the epoch
machinery bloated the service.

New files:
- extensionGalleryAccess.ts: shared leaf contracts (IExtensionGalleryAccessCore,
  IExtensionGalleryAccessProvider, IExtensionGalleryAccessSink, ICachedAccess,
  AccountResolution, ExtensionGalleryAccessProviderId, isSafeTokenTarget).
- extensionGalleryAccessProviders.ts: GitHub and Microsoft access providers.
- extensionGalleryAccessValidator.ts: provider-agnostic orchestrator.

Security invariants preserved: no microsoft->github fallback, cache scoped to
provider+serviceUrl, bearer only over HTTPS same-origin with followRedirects:0,
and the 401/403/transient status mappings are unchanged.

Copilot-Session: fe244b45-6362-4563-880d-fa3d9a27a07c
